+++ cmd/gokart/testdata/golden/structured-postgres/internal/commands/root.go
@@ -9,7 +9,7 @@ import (
 // Execute runs the CLI application.
 func Execute(version string) error {
 	cliApp := cli.NewApp("demo", version).
-		WithDescription("demo - a GoKart CLI application").
+		WithDescription("demo CLI").
 		WithEnvPrefix("DEMO").
 		WithStandardFlags()