+++ cmd/gokart/testdata/golden/structured/internal/commands/root.go
@@ -7,7 +7,7 @@ import (
 // Execute runs the CLI application.
 func Execute(version string) error {
 	cliApp := cli.NewApp("demo", version).
-		WithDescription("demo - a GoKart CLI application")
+		WithDescription("demo CLI")
 	cliApp.AddCommand(NewGreetCmd())
 
 	return cliApp.Run()